requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0 
langchain-text-splitters 
langchain-community
langchain-openai
langchain-chroma
pypdf 
chromadb 
jq
sentence-transformers
